FFmpeg target supports float frame rates

Closes #191

See merge request !21

+++ src/ffmpeg/ffmpeg_video_target.h
@@ -96,7 +96,7 @@ protected:
     //!
     //! \brief
     //!
-    int _framerate;
+    float _framerate;
 
     //!
     //! \brief Packet for use when encoding and
